修复orderby占位符解析和SQL验证逻辑
- 修复占位符解析：orderby/limit/offset占位符的完整内容作为options
- 修复SQL验证：nullable参数在验证前检查是否为空
- 修复TodoWebApi：CountAsync返回类型转换为int
- 修复测试脚本：404响应是预期行为
- 所有TodoWebApi API测试通过(17/17)

+++ samples/TodoWebApi/Program.cs
@@ -142,7 +142,7 @@ app.MapGet("/api/todos/due-soon", async (ITodoRepository repo) =>
 
 // Get total count
 app.MapGet("/api/todos/count", async (ITodoRepository repo) =>
-    Results.Json(await repo.CountAsync(), TodoJsonContext.Default.Int32));
+    Results.Json((int)await repo.CountAsync(), TodoJsonContext.Default.Int32));
 
 // Initialize database
 using var scope = app.Services.CreateScope();
